Refactor: extract DenyAction into its own component, add dedicated tests
- Extract deny button, confirmation dialog, and denied state into
  DenyAction component (one component per file per guidelines)
- ApprovalDetailScreen is now ~380 lines (down from ~610), focused on
  display; DenyAction handles the interaction lifecycle
- Add 8 dedicated DenyAction tests including auto-navigate timer,
  timer cleanup on unmount, and back-to-list button
- Simplify ApprovalDetailScreen tests to focus on layout/visibility
