+++ mobile/src/screens/approvals/__tests__/ApprovalDetailScreen.test.tsx
@@ -1,5 +1,4 @@
 import React, { createElement } from "react";
-import { Alert } from "react-native";
 import { create, act, type ReactTestRenderer } from "react-test-renderer";
 import type { ApprovalSummary } from "../../../hooks/useApprovals";
 import { makeApproval, MOCK_AGENTS, mockGetAgentDisplayName } from "../testFixtures";
@@ -7,12 +6,12 @@ import { makeApproval, MOCK_AGENTS, mockGetAgentDisplayName } from "../testFixtu
 // --- Mocks ---
 
 const mockDenyApproval = jest.fn();
-let mockIsDenying = false;
 
 jest.mock("../../../hooks/useDenyApproval", () => ({
   useDenyApproval: () => ({
     denyApproval: mockDenyApproval,
-    isPending: mockIsDenying,
+    isPending: false,
+    error: null,
   }),
 }));
 
@@ -57,7 +56,10 @@ function renderDetail(approval: ApprovalSummary) {
     key: "test",
     name: "ApprovalDetail" as const,
   };
-  const navigation = {} as any;
+  const navigation = {
+    canGoBack: jest.fn().mockReturnValue(true),
+    goBack: jest.fn(),
+  } as any;
 
   return create(
     createElement(ApprovalDetailScreen, { route, navigation } as any),
@@ -69,14 +71,6 @@ function hasDenyButton(renderer: ReactTestRenderer): boolean {
   return json.includes('"testID":"deny-button"');
 }
 
-function findDenyButton(renderer: ReactTestRenderer) {
-  // Pressable creates multiple fiber nodes; find the one with onPress
-  const nodes = renderer.root.findAll(
-    (node) => node.props.testID === "deny-button" && typeof node.props.onPress === "function",
-  );
-  return nodes[0];
-}
-
 // --- Tests ---
 
 describe("ApprovalDetailScreen", () => {
@@ -85,7 +79,6 @@ describe("ApprovalDetailScreen", () => {
   beforeEach(() => {
     jest.useFakeTimers();
     mockDenyApproval.mockReset();
-    mockIsDenying = false;
   });
 
   afterEach(async () => {
@@ -197,7 +190,7 @@ describe("ApprovalDetailScreen", () => {
     expect(json).toContain("Agent 999");
   });
 
-  // --- Deny flow tests ---
+  // --- Deny action visibility ---
 
   it("shows deny button for pending approval", async () => {
     const approval = makeApproval({ status: "pending" });
@@ -238,118 +231,5 @@ describe("ApprovalDetailScreen", () => {
       renderer = renderDetail(approval);
     });
     expect(hasDenyButton(renderer)).toBe(false);
-  });
-
-  it("shows confirmation alert when deny button is pressed", async () => {
-    const alertSpy = jest.spyOn(Alert, "alert");
-    const approval = makeApproval({ status: "pending" });
-    await act(async () => {
-      renderer = renderDetail(approval);
-    });
-
-    const denyButton = findDenyButton(renderer);
-    await act(async () => {
-      denyButton!.props.onPress();
-    });
-
-    expect(alertSpy).toHaveBeenCalledWith(
-      "Deny Request",
-      "Are you sure you want to deny this request?",
-      expect.arrayContaining([
-        expect.objectContaining({ text: "Cancel", style: "cancel" }),
-        expect.objectContaining({ text: "Deny", style: "destructive" }),
-      ]),
-    );
-    alertSpy.mockRestore();
-  });
-
-  it("calls denyApproval when confirmation is accepted", async () => {
-    mockDenyApproval.mockResolvedValue(undefined);
-    const alertSpy = jest.spyOn(Alert, "alert");
-    const approval = makeApproval({ status: "pending" });
-
-    await act(async () => {
-      renderer = renderDetail(approval);
-    });
-
-    const denyButton = findDenyButton(renderer);
-    await act(async () => {
-      denyButton!.props.onPress();
-    });
-
-    // Get the "Deny" button from the alert and press it
-    const alertButtons = alertSpy.mock.calls[0]![2] as Array<{ text: string; onPress?: () => void }>;
-    const denyAlertButton = alertButtons.find((b) => b.text === "Deny");
-
-    await act(async () => {
-      await denyAlertButton!.onPress!();
-    });
-
-    expect(mockDenyApproval).toHaveBeenCalledWith("appr_test123");
-    alertSpy.mockRestore();
-  });
-
-  it("shows denied banner after successful deny", async () => {
-    mockDenyApproval.mockResolvedValue(undefined);
-    const alertSpy = jest.spyOn(Alert, "alert");
-    const approval = makeApproval({ status: "pending" });
-
-    await act(async () => {
-      renderer = renderDetail(approval);
-    });
-
-    const denyButton = findDenyButton(renderer);
-    await act(async () => {
-      denyButton!.props.onPress();
-    });
-
-    const alertButtons = alertSpy.mock.calls[0]![2] as Array<{ text: string; onPress?: () => void }>;
-    const denyAlertButton = alertButtons.find((b) => b.text === "Deny");
-
-    await act(async () => {
-      await denyAlertButton!.onPress!();
-    });
-
-    const json = JSON.stringify(renderer.toJSON());
-    expect(json).toContain("Denied");
-
-    // Deny button should be gone after denial
-    expect(hasDenyButton(renderer)).toBe(false);
-
-    alertSpy.mockRestore();
-  });
-
-  it("shows error alert when deny fails", async () => {
-    mockDenyApproval.mockRejectedValue(new Error("Network error"));
-    const alertSpy = jest.spyOn(Alert, "alert");
-    const approval = makeApproval({ status: "pending" });
-
-    await act(async () => {
-      renderer = renderDetail(approval);
-    });
-
-    const denyButton = findDenyButton(renderer);
-    await act(async () => {
-      denyButton!.props.onPress();
-    });
-
-    // Press the "Deny" confirmation button
-    const alertButtons = alertSpy.mock.calls[0]![2] as Array<{ text: string; onPress?: () => void }>;
-    const denyAlertButton = alertButtons.find((b) => b.text === "Deny");
-
-    await act(async () => {
-      await denyAlertButton!.onPress!();
-    });
-
-    // Should show error alert
-    expect(alertSpy).toHaveBeenCalledWith(
-      "Error",
-      "Failed to deny request. Please try again.",
-    );
-
-    // Deny button should still be visible (not in denied state)
-    expect(hasDenyButton(renderer)).toBe(true);
-
-    alertSpy.mockRestore();
   });
 });
